AGClient: faking outgoing call indicators during SLC active call
When AG device gets connected to client device. during already
ongoing active call with it, we cant get info whether the call
is an outgoing call or incoming call.
So, always faking the call setup indicators as outgoing indicators
because, ring sound is heard if incoming indicators is sent.

CRs-Fixed: 4328135
Change-Id: I8d7e42f0cb564c5c4b90398233df0956334cf4ff