HFP: Suspend A2dp and Le Audio streaming when call update is received.
Even when no remote device is connected, when a call is received,
we need to send audio manager for LeAudio and A2dp suspend.
in cases like where only a2dp device is connected or in cases like
when HFP client is connected in concurrent call sharing feature,
we need to send the suspend params on receiving the call

Added change to send even though if no active device is connected.

CRs-Fixed: 4056868
Change-Id: Ic4182564276f31759fa86aea735962be02fd97e7