Pts: Forcing the stream start even if remote doesnt support the usecase context.
If remote is not supporting the audio context of the ongoing usecase,
we are not sending the stream start. For the PTS, based on the tc, remote
sometimes are not supporting the unspecified context support even.
In those cases, we are forcing the remote available context to start
streaming.
Added PTS based property check for forcing this.

CRs-Fixed: 4095484
Change-Id: Iff71aee956090b806387654969c09fdb9354be51