Set CONVERSATIONAL for bi-directional local metadata when call active
Issue case: make a call, during recording, call is mute.

Root cause: For this scenario, only local sink was updated
to CONVERSATIONAL when MM updated SRC metadata, the other
kept in LIVE which caused it changed to LIVE when re-align
contexts.

Fixed: Assign CONVERSATIONAL to bi-directional local metadata
when call is active.

Change-Id: Ia9d2c57ad9f996e4f9dc335302092ee28a789712
CRs-Fixed: 4212769