Enhancement for LE-A -> HFP SHO during music streaming.
Issue scenario:
1. Connect HFP watch and LE-A headset(active).
2. Music is streaming via LE-A HS.
3. Initial an outgoing call via HFP watch.
result: SCO is not setup, voice mute. After ending call,
music doesn't resume via LE-A HS.

Root cause: The call reconfiguration is not supposed to do
when device is under inactive. This reconfiguration won't
be completed once it starts, stack will block in releasing.
So that sco won't establish.

Fixed:
1. Ignore call reconfiguration when device is under inactive.
2. The phone state still in idle state when dialing outgoing call,
it needs to call setInactiveForHfpHandover to cache LE-A fallback
device.

Change-Id: Id4bfb6a621b64bc30b06928641e5fe8b2e8d320e
CRs-Fixed: 4243402