BCS: Fix: Prevent NullPointerException in BleConnectionViewModel
This change addresses a potential NullPointerException in the
`connectGattByScanning` method of the `BleConnectionViewModel` class.

If Bluetooth is turned off, `mBluetoothAdapter.getBluetoothLeScanner()`
can return null. This would cause a `NullPointerException` when
`bluetoothLeScanner` is subsequently used.

This patch adds a null check for the `BluetoothLeScanner` object. If it
is null, the method will now log a user-friendly message ("Please turn
on Bluetooth to use this App") and return gracefully.

CRs-Fixed: 4353947

Change-Id: Icf390a145ee205a8b223c81dfaf6d2221493aaba